refactor: frozenRowCell 重命名为 rowCell (#2450)
* refactor: frozenRowCell 重命名为 rowCell

* chore: 类型修改

* fix: 修复行头滚动刷选获取到错误的实例
