+++ packages/s2-core/__tests__/unit/utils/interaction/state-controller-spec.ts
@@ -1,5 +1,5 @@
+import type { RowCell } from '../../../../src/cell';
 import { getCellMeta } from '@/utils/interaction/select-event';
-import type { RowCell } from '@/cell/row-cell';
 import { CellTypes, InteractionStateName } from '@/common/constant/interaction';
 import type { S2Options } from '@/common/interface';
 import { Store } from '@/common/store';
