+++ packages/s2-core/src/cell/index.ts
@@ -4,12 +4,12 @@ import { CornerCell } from './corner-cell';
 import { DataCell } from './data-cell';
 import { HeaderCell } from './header-cell';
 import { MergedCell } from './merged-cell';
-import { RowCell } from './row-cell';
 import { TableColCell } from './table-col-cell';
 import { TableCornerCell } from './table-corner-cell';
 import { TableDataCell } from './table-data-cell';
 import { TableSeriesCell } from './table-series-cell';
-import { FrozenRowCell } from './frozen-row-cell';
+import { BaseRowCell } from './base-row-cell';
+import { RowCell } from './row-cell';
 import { SeriesNumberCell } from './series-number-cell';
 
 export {
@@ -18,12 +18,12 @@ export {
   TableSeriesCell,
   TableDataCell,
   RowCell,
+  BaseRowCell,
   ColCell,
   DataCell,
   MergedCell,
   CornerCell,
   BaseCell,
   HeaderCell,
-  FrozenRowCell,
   SeriesNumberCell,
 };
