+++ packages/s2-core/src/facet/header/pivot-row.ts
@@ -1,14 +1,14 @@
 import type { Node } from '../layout/node';
-import { FrozenRowCell, RowCell } from '../../cell';
+import { RowCell } from '../../cell/row-cell';
 import { BaseFrozenRowHeader } from './base-frozen-row';
 
 export class PivotRowHeader extends BaseFrozenRowHeader {
-  protected createCellInstance(item: Node): RowCell {
+  public createCellInstance(node: Node): RowCell {
     const { spreadsheet, scrollY } = this.headerConfig;
-    const frozenRow = this.isFrozenRow(item);
+    const frozenRow = this.isFrozenRow(node);
 
-    return new FrozenRowCell(
-      item,
+    return new RowCell(
+      node,
       spreadsheet,
       {
         ...this.headerConfig,
