+++ packages/s2-core/src/interaction/base-interaction/click/data-cell-click.ts
@@ -1,7 +1,6 @@
 import type { Event as CanvasEvent } from '@antv/g-canvas';
 import { forEach } from 'lodash';
 import type { DataCell } from '../../../cell/data-cell';
-import type { RowCell } from '../../../cell/row-cell';
 import {
   InteractionStateName,
   InterceptType,
