+++ packages/s2-react/__tests__/data/strategy-data.ts
@@ -389,6 +389,7 @@ export const StrategyOptions: SheetComponentOptions = {
   width: 800,
   height: 800,
   cornerText: '指标',
+  // frozenFirstRow: true,
   placeholder: (v) => {
     const placeholder = v?.fieldValue ? '-' : '';
     return placeholder;
