+++ packages/s2-react/playground/config.ts
@@ -80,7 +80,7 @@ export const s2Options: SheetComponentOptions = {
   debug: true,
   width: 600,
   height: 400,
-  frozenFirstRow: true,
+  frozenFirstRow: false,
   showSeriesNumber: false,
   interaction: {
     enableCopy: true,
