+++ s2-site/examples/basic/pivot/demo/grid.ts
@@ -10,7 +10,10 @@ fetch(
     const s2Options = {
       width: 600,
       height: 480,
+      // 冻结行头
+      // frozenRowHeader: true
     };
+
     const s2 = new PivotSheet(container, dataCfg, s2Options);
 
     s2.render();
